+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * change ownership of `/etc/profile.d/bpm.sh` to `vcap` group
 * improved consistency of error messages
 * add `/sbin` to the default system mounts
+* add support for privileged containers
 
 # v0.5.0